Move testing convenience fn to utils
